Intro to Jquery, and DOM explanation.
An HTML document is structured according to the Document Object Model,
or DOM. It's by interacting with the DOM that jQuery is able to access
and modify HTML.

The DOM consists of every element on the page, laid out in a
hierarchical way that reflects the way the HTML document is ordered.
Remember how we could think of the HTML document as a tree? You can
think of the DOM the same way. Just as with an HTML document, elements
in the DOM can have parents, children, and siblings
NOTE: the html has added in a new script link, just like css has to be
added in to the html doc.