Function breakdown part 2
A function is made up of three parts: the function keyword, any inputs
that function takes (they go between the ()s and are separated by
commas if there are more than one), and whatever actions the function
should perform (these go between the {}s). The general form is:

function(input1, input2, etc) {
    Do a thing
    Do another thing
    Do yet another thing!
}
One of the nice things about jQuery is that you can give a function
just about anything as an input—even another function! That's why
.ready() can take function between its parentheses; it's taking a
function as input