Variables and nth child selector
Variables are a place for us to store information for use at a later
time. Variables can hold any type of information you work with, so just
think of them as containers.

The single = sign is used to assign values. For instance, in jQuery,
you can write

var lucky = 7;
var name = "Codecademy";
var $p = $('p');

Our first variable contains a number, 7, while the second variable
contains some text, "Codecademy". Our 3rd variable stores the result of
a jQuery selector $('p') in the variable $p . As you can see, this is
just a handy way to save this information for later.

Why would we do this? Well, up until now we haven't had to modify
anything more than once. If we wanted to change the webpage based on
new information, we would need to store that information in variables.
Maybe you want to create a loading page and have it vanish as you
receive that information. It'd be a good idea to use variables.