Double selectors
Anything you can target with CSS, you can modify with jQuery. For
example, we can apply a fadeTo() to a p selector like this:

$('p').fadeTo('slow', 0);
We can apply a fadeTo() to an li selector like this:

$('li').fadeTo('slow', 0);
We can apply a fadeTo() to both the p and li selectors like this:

$('p, li').fadeTo('slow', 0);
This is called a compound selector.

Let's use a compound selector to apply a fadeTo() to both the .pink and
.red selectors.