This - to be used with an event handler ().fadeOut
In the last lesson, we had some code that looked like this:

$(document).ready(function() {
    $('div').mouseenter(function() {
        $('div').hide();
    });
});
The second line is good: this tells us that when we mouse into a div,
we should take a certain action. However, $('div').hide(); won't just
hide the div you mouse into; it will hide all the divs on the page. How
can we tell jQuery we only want to affect this particular div?

With this, of course!

The this keyword refers to the jQuery object you're currently doing
something with. Its complete rules are a little tricky, but the
important thing to understand is if you use an event handler on an
element—that's the fancy name for actions like .click() and
.mouseenter(), since they handle jQuery events—you can call the actual
event that occurs (such as fadeOut()) on $(this), and the event will
only affect the element you're currently doing something with (for
example, clicking on or mousing over).