Side down part 3
 we need to tell .click() what to do. In this case, when our pull tab
is clicked, we want our sliding panel (with the class .panel) to open
or close.

The jQuery event we need to toggle our sliding panel is (you guessed
it): .slideToggle()! We'll pass it one input, which is the speed of our
slide animation.

Inside .click()'s parentheses, create a .panel jQuery object and call
the .slideToggle() effect. Give it the input 'slow' for the animation
speed.