Append and prepend
.append() inserts the specified element as the last child of the target
element. .prepend() inserts the specified element as the first child of
the target element. If we have a div of class .info,

$(".info").append("<p>Stuff!</p>");
$(".info").prepend("<p>Stuff!</p>");
will add a paragraph containing the text "Stuff!" inside all divs of
class .info. .append() will make the paragraph the last child of each
div; .prepend() will make the paragraph the first child of each div.
(Note: see the Hint if you're using single quotes.)

.appendTo() does the same as .append(), but it just reverses the order
of "what to add" and "where to add it." The code

$('<p>Stuff!</p>').appendTo('.info');
has the same effect as the .append() code above. .prependTo() has a
similar relationship to .prepend().