To do list part 2
You'll notice we've set up an HTML form for grabbing the user's input.
We'll need to store the user's input in a variable, which will allow us
to append that input to the body of the HTML document later on.

You can set a variable equal to the contents of the input field using
.val(), like so:

//Get the value from an input
var input = $('input[name=checkListItem]').val();
Our selector finds our specific input using a css selector on our
checkListItem input
We call val() to get the value of the field
Instructions
Let's create a .click() event handler for our #button element
Make sure you pass it a function(){}
Inside your function(){} create the variable toAdd and set it equal to
the contents of the input field.