To do list part 3
Perfect! Now we want to add our HTML element to the document. We can do
this using our handy .append() function.

Let's go ahead and append to our div with the .list class. We'll append
a <div> with class="item", since we'll want a way to target our
appended <div>s later when we remove them. (A "to do" list is no good
if we can't check things off it.)

We'll want the contents of our div to be the contents of our input
field, which we saved in the variable toAdd. That means when we append,
we'll want to append

'<div class="item">' + toAdd + '</div>'
Instructions
Go ahead and .append() a <div> with class="item" to the .list div of
your HTML document, then MAKE SURE to click your button to add an
item—the exercise will wait for you to do so!

?
Hint
You'll need to alternate quotation marks here, since the HTML string
you're appending needs to be in quotes, but so does your class name.
The example in the instructional text above is one way to do this