Animate part 1
Now that you've got a solid foundation in jQuery events and have
covered important functions like .keydown() and .animate(), we'll put
our newfound knowledge to work. In this section, we'll animate a
sprite, or two-dimensional image on the screen.

For this momentous event, we've chosen a familiar face. Check him out
in the Result tab!