Animate part 3
Great work! Now it's time to animate our character based on the user's
input from the keyboard.

Every key press on a keyboard is translated into a number for the
computer to use. Don't worry about memorizing them, for now we've given
you the basics in script.js

// Left arrow key pressed
case 37:
  ('img').animate({left: "-=10px"}, 'fast');
The left arrow key on our keyboards translates to number 37 to the
computer. When that key is pressed, we animate our image to the left by
subtracting 10px
To move up we subtract 10px from the top
To move right we add 10px to the left
Finally, to move down we add 10px to the top
Can you fill in Up, Down, and Right? What happens if you add pixels
+=10px instead of subtracting?

Instructions
Let's make our image move!

Complete the code for Up, Right, and Down

The img needs to be clicked on before allowing the item to move.