Review part 2
Let's get in a little more practice with the .animate() effect. That's
the one with the slightly more complicated input—if we wanted to move a
div 10 pixels down, we'd type something like

$('div').animate({top:'+=10px'},500);
Where the bit between curly braces says "hey, jQuery! Add 10 pixels to
the current top margin," and the second input says "do it in 500
milliseconds!" (1,000 milliseconds = one second.)