Library example
The .effect() effect has all kinds of magical goodness in it, but it's
not the most amazing thing jQuery UI can do. The library has a number
of built-in effects that can make your website look sleek and
professional with surprisingly little code.

You can learn more in the jQuery UI documentation!

Instructions
Check out the code in script.js, then click Save & Submit Code and head
on over to the Result tab to see what jQuery UI can do for you. Be sure
to click each accordion section

http://jqueryui.com/