Selectable - uonordered lists
Images and <div>s aren't the only elements we can target with jQuery
UI—we can also enhance our ordered and unordered lists.

Check out the HTML in index.html. We've set up a pretty nifty ordered
list, but when it shows up in the Result tab, it's a little bland. With
jQuery UI, we can fix that!

Instructions
In the script.js tab, get your document .ready() and call the
.selectable() function on your 'ol'. Click Save & Submit Code, then
click on the list items in your ordered list in the Result tab to see
how much sleeker it looks.